+++ client/src/pages/auth/auth-layout.tsx
@@ -1,5 +1,7 @@
 import { Box, Text } from '@chakra-ui/react';
 import { Outlet } from 'react-router';
+
+import BetaDisclosure from '@/components/beta-disclosure';
 import { APP_CONFIG } from '@/config/app-config';
 
 const AuthLayout = () => {
@@ -9,8 +11,8 @@ const AuthLayout = () => {
       flexDirection="column"
       alignItems="stretch"
       justifyContent="center"
-      height="100vh"
-      gap={10}
+      minH="100vh"
+      gap={6}
       maxW="md"
       mx="auto"
       p={4}
@@ -18,6 +20,8 @@ const AuthLayout = () => {
       <Text fontSize="2xl" fontWeight="bold" textAlign="center">
         {APP_CONFIG.appName}
       </Text>
+
+      <BetaDisclosure contained={false} />
 
       <Outlet />
     </Box>